refactor: imports, exports, removed old hexagon, removed nested directories
