--- src/components/Hexagon/index.ts
@@ -1,1 +1,0 @@
-export * from './Hexagon';
