--- src/components/Hexagon/lookup.ts
@@ -1,44 +1,0 @@
-export const colors = {
-  primary: 'bg-primary',
-  secondary: 'bg-secondary',
-  tetriary: 'bg-tetriary',
-  quaternary: 'bg-quaternary',
-  pm: 'bg-guild-pm',
-  marketing: 'bg-guild-marketing',
-  webDev: 'bg-guild-webDev',
-  ux: 'bg-guild-ux',
-  dataScience: 'bg-guild-dataScience',
-  codersCamp: 'bg-guild-codersCamp',
-  white: 'bg-additional-white',
-  darkWhite: 'bg-additional-darkWhite'
-};
-
-export const dropShadowColors = {
-  primary: 'drop-shadow-[6px_8px_10px_#292929]',
-  secondary: 'drop-shadow-[6px_8px_10px_#4B545C]',
-  tetriary: 'drop-shadow-[6px_8px_10px_#9A9999]',
-  quaternary: 'drop-shadow-[6px_8px_10px_#1A90FF]',
-  pm: 'drop-shadow-[6px_8px_10px_#1BC55F]',
-  marketing: 'drop-shadow-[6px_8px_10px_#F0656F]',
-  webDev: 'drop-shadow-[6px_8px_10px_#0075FF]',
-  ux: 'drop-shadow-[6px_8px_10px_#FA8A43]',
-  dataScience: 'drop-shadow-[6px_8px_10px_#AC6DE8]',
-  codersCamp: 'drop-shadow-[6px_8px_10px_#FEC43D]',
-  white: 'drop-shadow-[6px_8px_10px_#FFFFFF]',
-  darkWhite: 'drop-shadow-[6px_8px_10px_#FAFAFA]'
-};
-
-export const opacities = {
-  0: 'bg-opacity-0',
-  5: 'bg-opacity-5',
-  10: 'bg-opacity-10',
-  20: 'bg-opacity-20',
-  30: 'bg-opacity-30',
-  40: 'bg-opacity-40',
-  50: 'bg-opacity-50',
-  60: 'bg-opacity-60',
-  70: 'bg-opacity-70',
-  80: 'bg-opacity-80',
-  90: 'bg-opacity-90',
-  100: 'bg-opacity-100'
-};
