+++ src/components/Hexagons/HexagonBase.tsx
