+++ src/components/Hexagons/IconHexagon.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
-import { HexagonBase } from '../HexagonBase/HexagonBase';
-import { variantMap } from '../lookup';
+import { HexagonBase } from './HexagonBase';
+import { variantMap } from './lookup';
 
 interface IconHexagonProps {
   variant: keyof typeof variantMap;
