+++ src/components/Hexagons/index.ts
@@ -1,0 +1,3 @@
+export * from './BackgroundHexagon';
+export * from './HexagonBase';
+export * from './IconHexagon';
