+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './Card';
 export * from './Carousel';
 export * from './Comment';
-export * from './Hexagon';
+export * from './Hexagons';
 export * from './Link';
 export * from './SectionWrapper';
